+++ docs/make.jl
@@ -96,6 +96,7 @@ const pages = [
             "Manual/assets/battery.md",
             "Manual/assets/beccselectricity.md",
             "Manual/assets/beccshydrogen.md",
+            "Manual/assets/constrainedfossilliquidfuels.md",
             "Manual/assets/downstreamemissions.md",
             "Manual/assets/electricdac.md",
             "Manual/assets/electricheating.md",