+++ src/MacroEnergy.jl
@@ -213,6 +213,7 @@ include("model/assets/hydrores.jl")
 include("model/assets/mustrun.jl")
 include("model/assets/upstreamemissions.jl")
 include("model/assets/downstreamemissions.jl")
+include("model/assets/constrainedfossilliquidfuels.jl")
 include("model/assets/syntheticnaturalgas.jl")
 include("model/assets/syntheticliquidfuels.jl")
 include("model/assets/syntheticammonia.jl")
@@ -285,6 +286,7 @@ export AbstractAsset,
     ElectricHeating,
     ElectricSteam,
     UpstreamEmissions,
+    ConstrainedFossilLiquidFuels,
     FuelCell,
     DownstreamEmissions,
     ThermalHeating,