+++ src/model/assets/electrolyzer.jl
@@ -1,7 +1,7 @@
 struct Electrolyzer <: AbstractAsset
     id::AssetId
     electrolyzer_transform::Transformation
-    h2_edge::Edge{<:Hydrogen}
+    h2_edge::Union{Edge{<:Hydrogen},EdgeWithUC{<:Hydrogen}}
     elec_edge::Edge{<:Electricity}
 end
 
@@ -53,6 +53,13 @@ function simple_default_data(::Type{Electrolyzer}, id=missing)
         :investment_cost => 0.0,
         :fixed_om_cost => 0.0,
         :variable_om_cost => 0.0,
+        :uc => false,
+        :startup_cost => 0.0,
+        :startup_fuel_consumption => 0.0,
+        :min_up_time => 0,
+        :min_down_time => 0,
+        :ramp_up_fraction => 1.0,
+        :ramp_down_fraction => 1.0,
     )
 end
 
@@ -85,7 +92,6 @@ end
 """
 function make(asset_type::Type{Electrolyzer}, data::AbstractDict{Symbol,Any}, system::System)
     id = AssetId(data[:id])
-    location = as_symbol_or_missing(get(data, :location, missing))
 
     @setup_data(asset_type, data, id)
 
@@ -103,7 +109,6 @@ function make(asset_type::Type{Electrolyzer}, data::AbstractDict{Symbol,Any}, sy
     electrolyzer = Transformation(;
         id = Symbol(id, "_", electrolyzer_key),
         timedata = system.time_data[Symbol(transform_data[:timedata])],
-        location = location,
         constraints = transform_data[:constraints],
     )
 
@@ -151,7 +156,9 @@ function make(asset_type::Type{Electrolyzer}, data::AbstractDict{Symbol,Any}, sy
         Hydrogen,
         [(h2_edge_data, :end_vertex), (data, :location)],
     )
-    h2_edge = Edge(
+    has_uc = get(h2_edge_data, :uc, false)
+    EdgeType = has_uc ? EdgeWithUC : Edge
+    h2_edge = EdgeType(
         Symbol(id, "_", h2_edge_key),
         h2_edge_data,
         system.time_data[:Hydrogen],
@@ -159,6 +166,18 @@ function make(asset_type::Type{Electrolyzer}, data::AbstractDict{Symbol,Any}, sy
         h2_start_node,
         h2_end_node,
     )
+    if has_uc
+        uc_constraints = [MinUpTimeConstraint(), MinDownTimeConstraint(), RampingLimitConstraint()]
+        for c in uc_constraints
+            if !(c in h2_edge.constraints)
+                push!(h2_edge.constraints, c)
+            end
+        end
+        # `startup_fuel_consumption` is specified in electricity units. The
+        # startup correction is applied to the hydrogen-output balance.
+        h2_edge.startup_fuel_consumption *= get(transform_data, :efficiency_rate, 1.0)
+        h2_edge.startup_fuel_balance_id = :energy
+    end
 
     @add_balance(
         electrolyzer,